+++ aptos-move/framework/aptos-stdlib/sources/cryptography/groth16.move
@@ -67,10 +67,10 @@ module aptos_std::groth16 {
     }
 
     native fun verify_proof_internal(
-        vk_alpha_g1_handle: u64, vk_beta_g_handle: u64, vk_gamma_g2_handle: u64, vk_delta_g2_handle: u64, gamma_abc_g1_handles: vector<u8>,
-        proof_a_handle: u64, proof_b_handle: u64, proof_c_handle: u64,
+        vk_alpha_g1_handle: u8, vk_beta_g_handle: u8, vk_gamma_g2_handle: u8, vk_delta_g2_handle: u8, gamma_abc_g1_handles: vector<u8>,
+        proof_a_handle: u8, proof_b_handle: u8, proof_c_handle: u8,
         public_input_handle: vector<u8>,
-        pairing_id: u64
+        pairing_id: u8
     ): bool;
 
     #[test]
